+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Picture type (I, P, B) information for HEVC streams in mp2ts-nallister
+- `-waitps` option to mp2ts-nallister to wait for parameter sets before printing NAL units
 - SMPTE-2038 data option to mp2ts-tools
 
 ## [0.2.1] - 2024-01-25